requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile
#
archivessnake==0.9.1
    # via -r requirements.in
asgiref==3.4.1
    # via django
asterism==0.7.2
    # via -r requirements.in
attrs==21.2.0
    # via
    #   archivessnake
    #   jsonschema
bagit==1.8.1
    # via
    #   -r requirements.in
    #   asterism
    #   bagit-profile
bagit-profile==1.3.1
    # via -r requirements.in
boltons==21.0.0
    # via archivessnake
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.3
    # via requests
clinner==1.12.3
    # via health-check
colorlog==3.2.0
    # via clinner
django==3.2.5
    # via
    #   -r requirements.in
    #   asterism
    #   django-braces
    #   django-common-helpers
    #   django-cron
    #   djangorestframework
django-braces==1.14.0
    # via -r requirements.in
django-common-helpers==0.9.2
    # via django-cron
django-cron==0.5.1
    # via -r requirements.in
django-datatables-view==1.19.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   asterism
djangorestframework-jwt==1.11.0
    # via -r requirements.in
gitdb==4.0.7
    # via gitpython
gitpython==3.1.18
    # via health-check
health-check==3.4.1
    # via -r requirements.in
idna==3.2
    # via requests
importlib-metadata==4.6.1
    # via jsonschema
iso-639==0.4.5
    # via -r requirements.in
iso8601==0.1.16
    # via -r requirements.in
jsonschema==3.2.0
    # via
    #   -r requirements.in
    #   rac-schemas
more-itertools==8.8.0
    # via archivessnake
mysqlclient==1.4.6
    # via -r requirements.in
odin==1.7.0
    # via asterism
psutil==5.8.0
    # via -r requirements.in
psycopg2-binary==2.9.1
    # via asterism
pyclamd==0.4.0
    # via -r requirements.in
pyjwt==1.7.1
    # via djangorestframework-jwt
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via
    #   -r requirements.in
    #   archivessnake
    #   health-check
rac_schemas==0.28
    # via -r requirements.in
rapidfuzz==1.4.1
    # via archivessnake
requests==2.26.0
    # via
    #   -r requirements.in
    #   archivessnake
    #   bagit-profile
six==1.16.0
    # via
    #   django-braces
    #   jsonschema
    #   odin
    #   python-dateutil
smmap==4.0.0
    # via gitdb
sqlparse==0.4.1
    # via django
structlog==21.1.0
    # via archivessnake
typing-extensions==3.10.0.0
    # via
    #   asgiref
    #   gitpython
    #   importlib-metadata
    #   structlog
uritemplate==3.0.1
    # via -r requirements.in
urllib3==1.26.6
    # via requests
zipp==3.5.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
